+++ test/form/samples/export-globals/_config.js
@@ -1,0 +1,6 @@
+module.exports = {
+	description: 'Supports reexports of globals with namespace access',
+	options: {
+		name: 'myBundle'
+	}
+};